==> 1.get_filers_from_year_city.ipynb <==
# Get state's filers

## Download the 2017 metadata and show where we put it

Use irsx_index at the command line to retrieve the 2018 and 2019 listing of all xml 990 efilings.

__Note that these are filings received *during* 2018/9, so check the tax_period__

Irsx_index is a helper command that is included with irsx, so you need to have installed it first. Try `pip install irsx` or see more [here](https://github.com/jsfenfen/990-xml-reader/#installation).

We use the `--verbose` flag so can watch it's progress

    $ irsx_index --verbose --year=2017
    Getting index file for year: 2017 
    remote=https://s3.amazonaws.com/irs-form-990/index_2017.csv 
    local=/Users/jfenton/github-whitelabel/envs/irs-cookbook/lib/python3.6/site-packages/irsx/CSV/index_2017.csv
    Beginning streaming download of https://s3.amazonaws.com/irs-form-990/index_2017.csv
    Total file size: 59.45 MB

## You don't need to run the command below if you've run irs_index at the command line
## To actually do this from within the notebook environment uncomment the below
## Note that we're using the %sx 'magic command' which captures the output as an array 
## Your mileage may vary depending on how jupyter plays with your operating system

## Getting 2018 and 2019 files here to ensure we've got the latest


%sx irsx_index --verbose --year=2018

%sx irsx_index --verbose --year=2019

# Get all nonprofit organizations in your state

We grabbed a file for just the state of Oregon as eo_or.csv from here: 
https://www.irs.gov/charities-non-profits/exempt-organizations-business-master-file-extract-eo-bmf

Note that this method isn't great for historic organizations; the IRS purges organizations after they've become inactive for a period of time. Historic EO BMF files are available here: http://nccs-data.urban.org/data.php?ds=bmf 

# importing libraries we'll use.
import csv
import os
import pandas as pd

# This tells us where the csv files are located in the system
from irsx.settings import INDEX_DIRECTORY

oregon_np = pd.read_csv("eo_or.csv")

# look at the first few lines
#oregon_np.head()
## print the headers as an array
list(oregon_np)

# Ignore some columns for now
or_np_simplified = oregon_np.filter(items=['EIN', 'NAME', 'ICO', 'STREET', 'CITY', 'STATE', 'ZIP', 'SUBSECTION','INCOME_AMT', 'ASSET_AMT', 'REVENUE_AMT', 'TAX_PERIOD', 'NTEE_CD'])
print("total oregon orgs: %s" % len(or_np_simplified))


# Show the top values by income 
or_np_simplified.sort_values(by=['INCOME_AMT'], ascending=[0]).head()



# this is from the index files we dowloaded at the start
INDEX_2018= os.path.join(INDEX_DIRECTORY, 'index_2018.csv')
np_2018 = pd.read_csv(INDEX_2018)

INDEX_2019= os.path.join(INDEX_DIRECTORY, 'index_2019.csv')
np_2019 = pd.read_csv(INDEX_2019)

#combine indexes and drop any older filings so we have latest filing for each org
df = pd.concat([np_2018,np_2019])
np_all = df.sort_values('TAX_PERIOD', ascending=False).drop_duplicates(subset=['EIN'])


## Now save the list of possible filers who actually filed out to .csv

# Now find orgs that are in oregon that filed in 2018 and 2019. 
# This join requires that both fields be named EIN and be formatted the same

ore_efilers = pd.merge(np_all,
                 or_np_simplified,
                 on='EIN')
print("Found a total of %s oregon 2018/19 efilers" % len(ore_efilers))
ore_efilers.head()

# sort by income amt, asset amt
ore_efilers = ore_efilers.sort_values(by=['INCOME_AMT', 'ASSET_AMT'], ascending=[0,0])
# Lets write them back out to a file for reference.
ore_efilers.to_csv('orefilers.csv')

# These are the top few for reference
ore_efilers.head()
