==> 2.2-ez_comp_output.ipynb <==
# Dump salaries from 990EZ in OR (continuing from part 1)

import unicodecsv as csv
from irsx.xmlrunner import XMLRunner
import pandas as pd

# filers = pd.read_csv("orefilers.csv")
# ez = filers["RETURN_TYPE"] == "990EZ"
# reader = filers[ez]
# reader.to_csv('orefilers_ez.csv',index=False)

# read the whole file of orgs with efilings from part 1 here, it's not very long
ez_rows = [] 
# We're using the output of part 1
with open('orefilers.csv', 'rb') as infile:
    reader = csv.DictReader(infile)
    for row in reader:
        if row['RETURN_TYPE'] == "990EZ" or row['RETURN_TYPE'] == "990EO":
            ez_rows.append(row['OBJECT_ID'])
ez_rows

# the name of the output file
outfilename ="ez_comp_1.csv"
outfile = open(outfilename , 'wb')

# the header rows as they'll appear in the output
headers = ["ein", "object_id", "exec_comp"]
# start up a dictwriter, ignore extra rows
dw = csv.DictWriter(outfile, headers, extrasaction='ignore')
dw.writeheader()

# get an XMLRunner -- this is what actually does the parsing
xml_runner = XMLRunner()

## Figure out what to extract

Data from each repeating group should go to it's own file, otherwise it won't make sense.

To figure out what to capture, I started by looking at schedule J: http://www.irsx.info/#IRS990ScheduleJ
Then I went to the table details and picked the rows I wanted from the repeating group:
http://www.irsx.info/metadata/groups/SkdJRltdOrgOffcrTrstKyEmpl.html

Note that it's common for director/employee names in schedule J to get listed as businessname.

Also note that IRSx checks to see if a file has been downloaded before fetching it. Running this the first time will be slow if the filings aren't already downloaded, but much faster if they've already been downloaded.

-------NOTE--------
Edited to get compensation from EZ partIV, not sked j


execs = {}
for object_id in file_rows:
    parsed_filing = xml_runner.run_filing(object_id)
    parsed_main = parsed_filing.get_parsed_sked('IRS990EZ')[0]
    employee_groups = parsed_main['groups']['EZOffcrDrctrTrstEmpl']
    exec_comp = 0
    for employee_group in employee_groups:

        comp = int(employee_group.get('CmpnstnAmt'))
        if comp>0:
            exec_comp += comp
        if employee_group.get('EmplyBnftPrgrmAmt'):
            exec_comp += int(employee_group.get('EmplyBnftPrgrmAmt'))
        if employee_group.get('ExpnsAccntOthrAllwncAmt'):
            exec_comp += int(employee_group.get('ExpnsAccntOthrAllwncAmt'))

    execs[object_id] = exec_comp

execs['201713319349200006']

DEMO_MAX = 40
num_rows = 0

all_rows = []

for row in file_rows:
    num_rows += 1
    this_object_id = row['OBJECT_ID']
    parsed_filing = xml_runner.run_filing(this_object_id)
    
    # if it somehow busted, just note it and continue
    if not parsed_filing:
        print("Skipping filing %s(filings with pre-2013 schemas are skipped)\n row details: %s" % (this_object_id, row))
        continue 
    
    schedule_list = parsed_filing.list_schedules()
    exec_comp = 0
    if 'IRS990EZ' in schedule_list:
        
        
        # store the output in this dict
        outputdata = {}
        # assign some initial values from the input csv
        outputdata['ein'] = row['EIN']
        outputdata['object_id'] = row['OBJECT_ID']
        
        parsed_main = parsed_filing.get_parsed_sked('IRS990EZ')[0] 
        
        # now use the table name we looked up -- that's where we find it
        try:
            # repeating groups are returned as an array of dicts
            employee_groups = parsed_main['groups']['EZOffcrDrctrTrstEmpl']
        except KeyError:
            print("No EZOffcrDrctrTrstEmpl found in %s skipping" % this_object_id)
            continue
          
        # read through each employee and pull out the data we want
        for employee_group in employee_groups:
            
            comp = int(employee_group.get('CmpnstnAmt'))
            if comp>0:
                exec_comp += comp
            if employee_group.get('EmplyBnftPrgrmAmt'):
                exec_comp += int(employee_group.get('EmplyBnftPrgrmAmt'))
            if employee_group.get('ExpnsAccntOthrAllwncAmt'):
                exec_comp += int(employee_group.get('ExpnsAccntOthrAllwncAmt'))

        outputdata['exec_comp'] = exec_comp
        all_rows.append(outputdata)
#         dw.writerow(outputdata)
        

# Don't run endlessly during a demo:
#     if(num_rows > DEMO_MAX):
#        break

#     if num_rows%100==0:
#         print("Processed %s filings" % num_rows)
        
# close the outfile
# outfile.close()

# print("end")

s = next((item for item in all_rows if item["ein"] == "814325149"), False)
print(s['exec_comp'])
